% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.biomimic_fit}
\alias{print.biomimic_fit}
\title{Print Method for biomimic_fit}
\usage{
\method{print}{biomimic_fit}(x, ...)
}
\arguments{
\item{x}{A \code{biomimic_fit} object.}

\item{...}{Ignored.}
}
\description{
Print Method for biomimic_fit
}
