% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{print.biomimic_gamma_calibration}
\alias{print.biomimic_gamma_calibration}
\title{Print method for biomimic_gamma_calibration objects}
\usage{
\method{print}{biomimic_gamma_calibration}(x, ...)
}
\arguments{
\item{x}{A \code{biomimic_gamma_calibration} object.}

\item{...}{Ignored.}
}
\description{
Print method for biomimic_gamma_calibration objects
}
