% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{print.biomimic_gof}
\alias{print.biomimic_gof}
\title{Print Method for biomimic_gof}
\usage{
\method{print}{biomimic_gof}(x, ...)
}
\arguments{
\item{x}{A \code{biomimic_gof} object.}

\item{...}{Ignored.}
}
\description{
Print Method for biomimic_gof
}
