% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invariance.R
\name{print.biomimic_invariance}
\alias{print.biomimic_invariance}
\title{Print Method for biomimic_invariance}
\usage{
\method{print}{biomimic_invariance}(x, ...)
}
\arguments{
\item{x}{A \code{biomimic_invariance} object.}

\item{...}{Ignored.}
}
\description{
Print Method for biomimic_invariance
}
