% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.biomimic_selection}
\alias{print.biomimic_selection}
\title{Print method for biomimic_selection objects}
\usage{
\method{print}{biomimic_selection}(x, ...)
}
\arguments{
\item{x}{A \code{biomimic_selection} object.}

\item{...}{Ignored.}
}
\description{
Print method for biomimic_selection objects
}
