% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.biomimic_summary}
\alias{print.biomimic_summary}
\title{Print Method for biomimic_summary}
\usage{
\method{print}{biomimic_summary}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{biomimic_summary} object.}

\item{digits}{Number of decimal places (default 3).}

\item{...}{Ignored.}
}
\description{
Print Method for biomimic_summary
}
