% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.vb_ard}
\alias{print.vb_ard}
\title{Print method for vb_ard objects}
\usage{
\method{print}{vb_ard}(x, ...)
}
\arguments{
\item{x}{A \code{vb_ard} object.}

\item{...}{Ignored.}
}
\description{
Print method for vb_ard objects
}
